requirements.txt
# For neural network examples
torch
torchvision

# For declarative query 
pyyaml 